backend/requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0
SQLAlchemy==2.0.37
alembic==1.14.1
psycopg[binary]==3.2.4
pydantic==2.10.6
pydantic-settings==2.7.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
email-validator==2.2.0
python-multipart==0.0.20
boto3==1.36.16
pytest==8.3.4
hypothesis==6.125.2
httpx==0.28.1
